feat: update header table component
